+++ SampleCode/app/src/main/java/com/pingidentity/sdk/pingonewallet/sample/callbacks/DocumentClickListener.java
@@ -1,9 +1,0 @@
-package com.pingidentity.sdk.pingonewallet.sample.callbacks;
-
-import com.pingidentity.sdk.pingonewallet.sample.models.Credential;
-
-public interface DocumentClickListener {
-
-    void onActionClick(Credential credential);
-
-}