+++ SampleCode/app/src/main/java/com/pingidentity/sdk/pingonewallet/sample/rv_adapters/view_holders/CardViewHolder.java
@@ -6,9 +6,10 @@ import android.view.View;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.pingidentity.sdk.pingonewallet.sample.databinding.ItemCredentialCardBinding;
-import com.pingidentity.sdk.pingonewallet.sample.callbacks.DocumentClickListener;
 import com.pingidentity.sdk.pingonewallet.sample.models.Credential;
 import com.pingidentity.sdk.pingonewallet.sample.utils.BitmapUtil;
+
+import java.util.function.Consumer;
 
 public class CardViewHolder extends RecyclerView.ViewHolder {
 
@@ -19,7 +20,7 @@ public class CardViewHolder extends RecyclerView.ViewHolder {
         this.mBinding = binding;
     }
 
-    public void bind(Credential credential, DocumentClickListener callback) {
+    public void bind(Credential credential, Consumer<Credential> callback) {
         String cardType = credential.getClaim().getData().getOrDefault("CardType", "");
         mBinding.txtCardTitle.setText(cardType);
         Bitmap image = BitmapUtil.getBitmapFromClaim(credential.getClaim());
@@ -30,7 +31,7 @@ public class CardViewHolder extends RecyclerView.ViewHolder {
             mBinding.cardFrontImage.setVisibility(View.GONE);
         }
         mBinding.viewExpired.setVisibility(credential.isRevoked() ? View.VISIBLE : View.GONE);
-        mBinding.layoutDocument.setOnClickListener(v -> callback.onActionClick(credential));
+        mBinding.layoutDocument.setOnClickListener(v -> callback.accept(credential));
     }
 
 }