+++ SampleCode/app/src/main/java/com/pingidentity/sdk/pingonewallet/sample/ui/picker/PickerViewModel.java
@@ -1,0 +1,13 @@
+package com.pingidentity.sdk.pingonewallet.sample.ui.picker;
+
+import com.pingidentity.sdk.pingonewallet.sample.wallet.PingOneWalletHelper;
+import com.pingidentity.sdk.pingonewallet.storage.data_repository.DataRepository;
+import com.pingidentity.sdk.pingonewallet.sample.ui.base.BaseViewModel;
+
+public class PickerViewModel extends BaseViewModel {
+
+    public PickerViewModel(PingOneWalletHelper pingOneWalletHelper) {
+        super(pingOneWalletHelper);
+    }
+
+}